+++ src/Rask.Server/RaskEndpointExtensions.cs
@@ -204,6 +204,10 @@ public static class RaskEndpointExtensions
         services.AddScoped<IBroadcastChannel, BroadcastChannelService>();
         services.AddScoped<IIntersectionObserver, IntersectionObserverService>();
         services.AddScoped<IResizeObserver, ResizeObserverService>();
+        services.AddScoped<IMutationObserver, MutationObserverService>();
+        services.AddScoped<IMediaSession, MediaSession>();
+        services.AddScoped<IDeviceOrientation, DeviceOrientation>();
+        services.AddScoped<IDeviceMotion, DeviceMotion>();
         services.AddScoped<ICrypto, Crypto>();
         services.AddScoped<IPerformance, Performance>();
         services.AddScoped<IIndexedDb, IndexedDb>();